Move mockAzureModuleData to mockData folder
