+++ src/client/src/mockData/extensionModules/mockAzureModule.ts
@@ -1,6 +1,6 @@
 import { EXTENSION_COMMANDS, EXTENSION_MODULES } from "../../utils/constants/commands";
 
-import * as mockData from "./mockAzureModuleData";
+import * as mockData from "./mockData/mockAzureModuleData";
 
 const DEV_NO_ERROR_MSG = "in development, no error message";
 const DEV_NO_ERROR_TYPE = "in development, no error type";
